requirements.txt
astroid==2.11.6
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
dill==0.3.5.1
filelock==3.7.1
gdown==4.4.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
platformdirs==2.5.2
pylint==2.14.3
PySocks==1.7.1
requests==2.28.0
scipy==1.8.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
